Add RH and LH Perspective and PrespectiveReversed
Also change the utility functions to use Float instead of Double

+++ Sources/SLA/SLA.swift
@@ -1,14 +1,14 @@
 import Foundation
 
-public let PI  = Double.pi
-public let TAU = Double.pi * 2.0
+public let PI  = Float.pi
+public let TAU = Float.pi * 2.0
 
-public func deg2rad(_ d: Double) -> Double
+public func deg2rad(_ d: Float) -> Float
 {
     return d * TAU / 360.0
 }
 
-public func rad2deg(_ r: Double) -> Double
+public func rad2deg(_ r: Float) -> Float
 {
     return r * 360.0 / TAU
 }
